requirements.txt
appdata==2.2.1
asgiref==3.7.2
click==8.0.3
colorama==0.4.6
colored==1.4.3
Django==5.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
to-requirements.txt==2.0.0
tzdata==2023.3
uritemplate==4.1.1
